+++ CHANGELOG.md
@@ -1,12 +1,20 @@
 # C++ Game Frame 32 Changelog
 
-## Version X.y.z [dd-mm-23]
+## Version 0.y.z [dd-mm-23]
 ### `**Version name**`
 ### Added
 ### Changed
 ### Fixed
 ### Deleted
-
-
+---
+## Version 0.0.1 [25-02-23]
+### `**Core functionality [WIP]**`
+### Added
+- `Engine32` class
+- `AsyncWebServer` communication
+- `Routes`:
+ >- `[GET] /` - returns index HTML with `m32.js` engine
+ >- TODO: `[GET] /events` - returns the current EventBuffer 
+---
 ## Version 0.0.0 [25-02-23]
 ### `**Initial Setup**`